requirement.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-import-export==3.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.1.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.2
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
random-user-agent==1.0.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
webdriver-manager==3.8.5
webencodings==0.5.1
whitenoise==6.4.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
